+++ controllers/event_controllers/checkIfEventExists.js
