+++ controllers/event_controllers/confirmEventStatus.js
