+++ routes/book.routes.js
@@ -2,8 +2,8 @@ const { Router } = require('express')
 
 const User = require('../models/User.model')
 const Book = require('../models/Book.model')
-const { createBookReqPayload } = require('./book_functions/createBookReqPayload')
-const { verifyUserId } = require('./helper_functions/verifyUserId.js')
+const { createBookReqPayload } = require('../controllers/book_controllers/createBookReqPayload')
+const { verifyUserId } = require('../controllers/helper_controllers/verifyUserId.js')
 
 const router = Router()
